Move ProjectDeadline enum to project entity namespace
